refactor application structure
