+++ backend/src/main/java/com/example/model/Platform.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.model;
 
 import lombok.Data;
