+++ backend/src/test/java/com/example/AllocationControllerBestRateTest.java
@@ -1,5 +1,6 @@
 package com.example;
 
+import com.example.model.Allocation;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
