+++ backend/src/test/java/com/example/AllocationControllerTest.java
@@ -1,11 +1,9 @@
 package com.example;
 
+import com.example.model.Platform;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
-
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AllocationControllerTest {
