+++ frontend/src/App.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import {useSelector} from "react-redux";
-import {Card} from "./components/presentation/Panels";
-import {FEATURES} from "./features";
-import {BestRate} from "./components/container/BestRate";
-import {AllocationsCalculator} from "./components/container/AllocationsCalculator";
-import {VerticalContainer} from "./components/presentation/VerticalContainer";
+import {Card} from "./presentation/Panels";
+import {FEATURES} from "./domain/features/features";
+import {BestRate} from "./domain/best_rate/BestRate";
+import {AllocationsCalculator} from "./domain/allocations/AllocationsCalculator";
+import {VerticalContainer} from "./presentation/VerticalContainer";
 
 export const App = () => {
   const features = useSelector(x => x.features)
