+++ frontend/src/domain/allocations/constants.js
