+++ frontend/src/domain/best_rate/BestRate.jsx
@@ -1,14 +1,14 @@
 import {useDispatch, useSelector} from "react-redux";
 import React, {useEffect} from "react";
-import {bestRateFetched} from "../../redux/actions/allocations";
-import {Card} from "../presentation/Panels";
-import {formatRate} from "../../utils/formatting";
-import {fetchJson} from "../../utils/request";
-import {errorsAdded} from "../../redux/actions/errors";
+import {bestRateFetched} from "../allocations/actions";
+import {Card} from "../../presentation/Panels";
+import {formatRate} from "../../presentation/formatting";
+import {fetchJson} from "../../http/request";
+import {errorsAdded} from "../errors/actions";
 
 export const BestRate = () => {
   const dispatch = useDispatch()
-  
+
   const bestAllocation = useSelector(x => x.allocations.bestRate)
 
   useEffect(() => {
