+++ frontend/src/domain/best_rate/BestRate.spec.js
@@ -2,11 +2,11 @@ import nock from 'nock'
 import {render, waitFor} from "@testing-library/react";
 import React from "react";
 import {Provider} from "react-redux";
-import {store} from "../../redux/redux";
+import {store} from "../../redux";
 import 'whatwg-fetch' // sets global.fetch
 import '@testing-library/jest-dom'
 import {BestRate} from "./BestRate";
-import {formatRate} from "../../utils/formatting";
+import {formatRate} from "../../presentation/formatting";
 
 function getBestRateSurroundingDivId() {
   return 'allocation-c020b901';
