+++ frontend/src/domain/errors/ErrorAlert.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import {useDispatch, useSelector} from "react-redux";
 import * as r from 'ramda'
-import {errorsDismissedFirst} from "../../redux/actions/errors";
-import {ErrorAlertPresentation} from "../presentation/ErrorAlertPresentation";
+import {errorsDismissedFirst} from "./actions";
+import {ErrorAlertPresentation} from "./ErrorAlertPresentation";
 
 export const ErrorAlert = () => {
   const errors = useSelector(x => x.errors)
