+++ frontend/src/domain/errors/actions.js
@@ -1,4 +1,4 @@
-import {ERRORS_ADDED, ERRORS_DISMISSED_FIRST} from "../constants/errors";
+import {ERRORS_ADDED, ERRORS_DISMISSED_FIRST} from "./constants";
 
 export const errorsDismissedFirst = () => ({type: ERRORS_DISMISSED_FIRST})
 export const errorsAdded = (msg) => ({type: ERRORS_ADDED, payload: msg})
