+++ frontend/src/domain/errors/constants.js
