+++ frontend/src/domain/features/actions.js
@@ -1,3 +1,3 @@
-import {FEATURE_CHANGED} from "../constants/features";
+import {FEATURE_CHANGED} from "./constants";
 
 export const featureChanged = (name, value) => ({type: FEATURE_CHANGED, payload: {[name]: value}})
