+++ frontend/src/domain/features/features.js
@@ -1,6 +1,6 @@
 import {SplitFactory} from "@splitsoftware/splitio";
-import {store} from "./redux/redux";
-import {featureChanged} from "./redux/actions/features";
+import {store} from "../../redux";
+import {featureChanged} from "./actions";
 
 export const FEATURES = {
   FIRST: "my-first-split",
