+++ frontend/src/domain/features/reducers.js
@@ -1,5 +1,5 @@
 import * as r from "ramda"
-import {FEATURE_CHANGED} from "../constants/features";
+import {FEATURE_CHANGED} from "./constants";
 
 export const featuresReducer = (state = {}, action) => {
   switch (action.type) {
