+++ frontend/src/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {App} from "./App";
-import './features'
-import './redux/redux'
+import './domain/features/features'
+import './redux'
 import {Provider} from "react-redux";
-import {store} from "./redux/redux";
+import {store} from "./redux";
 import {AppShell} from "./AppShell";
 
 const app = (
