+++ frontend/src/presentation/Panels.jsx
