+++ frontend/src/presentation/Panels.spec.js
