+++ frontend/src/presentation/formatting.js
