+++ frontend/src/presentation/formatting.spec.js
